refactor: restructure project packages
