+++ src/main/java/gift/exception/GlobalExceptionHandler.java
@@ -1,6 +1,5 @@
-package gift;
+package gift.exception;
 
-import org.springframework.dao.DuplicateKeyException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
